requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
mutagen==1.47.0
ffmpeg-python==0.2.0